backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3

fpdf2==2.7.6
python-docx>=0.8.11
aiofiles==23.2.1
transformers>=4.36.0
torch
spacy>=3.7.0
pytest>=7.0.0
httpx==0.27.0
sqlalchemy>=2.0.0
faiss-cpu
sentence-transformers
passlib[bcrypt]
PyJWT

# Security & Configuration
pydantic-settings>=2.0.0
slowapi>=0.1.9

# OCR & Translation
pytesseract
pdf2image
Pillow
deep-translator

# LLM Providers (install at least one)
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
